+++ src/Rule/Type/BooleanType.php
@@ -1,0 +1,93 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Validator\Rule\Type;
+
+use Attribute;
+use Closure;
+use Yiisoft\Validator\Rule\Trait\SkipOnEmptyTrait;
+use Yiisoft\Validator\Rule\Trait\SkipOnErrorTrait;
+use Yiisoft\Validator\Rule\Trait\WhenTrait;
+use Yiisoft\Validator\RuleWithOptionsInterface;
+use Yiisoft\Validator\SkipOnEmptyInterface;
+use Yiisoft\Validator\SkipOnErrorInterface;
+use Yiisoft\Validator\WhenInterface;
+
+/**
+ * Defines validation options to check that the value has boolean type. It's different from {@see BooleanValue} because
+ * it only strictly allows boolean values ({@link https://www.php.net/manual/en/language.types.boolean.php}) and not the
+ * ones that evaluates to boolean ("truthy" / "falsy"). If you need to support that, consider to use {@see BooleanValue}
+ * instead.
+ *
+ * @see BooleanTypeHandler Corresponding handler performing the actual validation.
+ *
+ * @psalm-import-type SkipOnEmptyValue from SkipOnEmptyInterface
+ * @psalm-import-type WhenType from WhenInterface
+ */
+#[Attribute(Attribute::TARGET_PROPERTY | Attribute::IS_REPEATABLE)]
+final class BooleanType implements RuleWithOptionsInterface, SkipOnEmptyInterface, SkipOnErrorInterface, WhenInterface
+{
+    use SkipOnEmptyTrait;
+    use SkipOnErrorTrait;
+    use WhenTrait;
+
+    /**
+     * @param string $message Error message used when validation fails because the type of the validated value is not
+     * boolean.
+     *
+     * You may use the following placeholders in the message:
+     *
+     * - `{attribute}`: the translated label of the attribute being validated.
+     * - `{type}`: the type of the value being validated.
+     * @param bool|callable|null $skipOnEmpty Whether to skip this rule if the validated value is empty / not passed.
+     * See {@see SkipOnEmptyInterface}.
+     * @param bool $skipOnError Whether to skip this rule if any of the previous rules gave an error. See
+     * {@see SkipOnErrorInterface}.
+     * @param Closure|null $when A callable to define a condition for applying the rule. See {@see WhenInterface}.
+     *
+     * @psalm-param SkipOnEmptyValue $skipOnEmpty
+     * @psalm-param WhenType $when
+     */
+    public function __construct(
+        private string $message = 'Value must be a boolean.',
+        private mixed $skipOnEmpty = null,
+        private bool $skipOnError = false,
+        private Closure|null $when = null,
+    ) {
+    }
+
+    public function getName(): string
+    {
+        return 'booleanType';
+    }
+
+    /**
+     * Gets error message used when validation fails because the type of validated value is not boolean.
+     *
+     * @return string Error message / template.
+     *
+     * @see $message
+     */
+    public function getMessage(): string
+    {
+        return $this->message;
+    }
+
+    public function getOptions(): array
+    {
+        return [
+            'message' => [
+                'template' => $this->message,
+                'parameters' => [],
+            ],
+            'skipOnEmpty' => $this->getSkipOnEmptyOption(),
+            'skipOnError' => $this->skipOnError,
+        ];
+    }
+
+    public function getHandler(): string
+    {
+        return BooleanTypeHandler::class;
+    }
+}